fix(conditional compilation): make linux-only
The conditional compilation in the crate was broken so I divided it up into a platform module instead and now it runs
only on Linux to be on the safe side. Should be pretty straightforward to add other platforms eventually
