+++ src/platform/mod.rs
@@ -1,0 +1,14 @@
+#[cfg(unix)]
+mod unix_pipe;
+
+#[cfg(target_os = "linux")]
+mod linux;
+
+#[cfg(not(target_os = "linux"))]
+mod unsupported;
+
+#[cfg(target_os = "linux")]
+pub use linux::*;
+
+#[cfg(not(target_os = "linux"))]
+pub use unsupported::*;
